# internal/docs/swagger.yaml
basePath: /api
definitions:
...
  participants.createParticipantRequest:
    properties:
      display_name:
        maxLength: 160
        minLength: 2
        type: string
      external_identifier:
        type: string
    required:
    - display_name
    type: object
  participants.participantResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      external_identifier:
        type: string
      id:
        type: string
      project_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  participants.updateParticipantStatusRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API with Clean Architecture
  title: Indagio API
  version: "1.0"
paths:
...
  /projects/{id}/participants:
    get:
      description: List participants for a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/participants.participantResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List participants
      tags:
      - Participants
    post:
      consumes:
      - application/json
      description: Create a participant in a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Participant payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/participants.createParticipantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/participants.participantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create participant
      tags:
      - Participants
...
  /projects/{id}/participants/{participantID}/status:
    patch:
      consumes:
      - application/json
      description: Update participant status (for example active, paused, archived)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Participant ID
        in: path
        name: participantID
        required: true
        type: string
      - description: Status payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/participants.updateParticipantStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/participants.participantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update participant status
      tags:
      - Participants
  /projects/{id}/participants/code/{code}:
    get:
      description: Get one participant by project-scoped code
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Participant code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/participants.participantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get participant by code
      tags:
      - Participants
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"